sample data
@2015characterlistprimarychinese.pdf contains a list of chinese words. are you able to construct a couple of simple singapore primary school chinese sentences for each word and write them into output.json containing []{ word, word_hanyupinyin, sentence, sentence_hanyupinyin }

sorry can we change the schema to add word_english and sentence_english too?

can we write a basic Go program to validate a given json file (of this schema) -- that the "sentence" contains the "word". i suspect there are invalid sentences in 2015characterlistprimarychinese.pdf.json that does not contain the word and would like a cli to highlight them

yes fix the 13 invalid sentences and rerun the command to check